+++ src/Data/Geo/Jord/Length.hs
@@ -11,7 +11,7 @@
 module Data.Geo.Jord.Length
     (
     -- * The 'Length' type
-      Length(millimetres)
+      Length
     -- * Smart constructors
     , feet
     , kilometres
@@ -36,9 +36,9 @@ import Prelude hiding (fail, length)
 import Text.ParserCombinators.ReadP
 import Text.Read hiding (pfail)
 
--- | A length with a resolution of 1 millimetre.
+-- | A length with a resolution of 0.1 millimetre.
 newtype Length = Length
-    { millimetres :: Int
+    { tenthOfMm :: Int
     } deriving (Eq)
 
 -- | See 'readLength'.
@@ -53,10 +53,10 @@ instance Show Length where
       where
         m = toMetres l
 
--- | Add/Subtract Length.
+-- | Add/Subtract 'Length's.
 instance Quantity Length where
-    add a b = Length (millimetres a + millimetres b)
-    sub a b = Length (millimetres a - millimetres b)
+    add a b = Length (tenthOfMm a + tenthOfMm b)
+    sub a b = Length (tenthOfMm a - tenthOfMm b)
     zero = Length 0
 
 -- | 'Length' from given amount of feet.
@@ -69,7 +69,7 @@ kilometres km = metres (km * 1000.0)
 
 -- | 'Length' from given amount of metres.
 metres :: Double -> Length
-metres m = Length (round (m * 1000.0))
+metres m = Length (round (m * 10000.0))
 
 -- | 'Length' from given amount of nautical miles.
 nauticalMiles :: Double -> Length
@@ -107,7 +107,7 @@ toKilometres l = toMetres l / 1000.0
 
 -- | @toMetres l@ converts @l@ to metres.
 toMetres :: Length -> Double
-toMetres (Length mm) = fromIntegral mm / 1000.0
+toMetres (Length mm) = fromIntegral mm / 10000.0
 
 -- | @toNauticalMiles l@ converts @l@ to nautical miles.
 toNauticalMiles :: Length -> Double