+++ test/Data/Geo/Jord/LengthSpec.hs
@@ -30,6 +30,11 @@ spec = do
             toKilometres (nauticalMiles 10.5) `shouldBe` 19.446
         it "converts feet to metres" $ toMetres (feet 25000) `shouldBe` 7620
         it "converts metres to feet" $ toFeet (metres 7620) `shouldBe` 25000
+    describe "Resolution" $ do
+        it "handles 1 kilometre" $ toKilometres (kilometres 1) `shouldBe` 1
+        it "handles 1 metre" $ toMetres (metres 1) `shouldBe` 1
+        it "handles 1 nautical mile" $ toNauticalMiles (nauticalMiles 1) `shouldBe` 1
+        it "handles 1 foot" $ toFeet (feet 1) `shouldBe` 1
     describe "Adding/Subtracting lengths" $ do
         it "adds lengths" $ add (kilometres 1000) (metres 1000) `shouldBe` metres 1001000
         it "subtracts lengths" $ sub (metres 1000) (nauticalMiles 10.5) `shouldBe` metres (-18446)